Topic/db/query improvements (#127)
* TextCategoryRepository.java: add @EntityGraph annotation on 'category' attribute to eager load association
called in TextcategoryService:loadTextCategoryIdMap()

* [api] application-test.properties: limit Hikari connection pool count when using pgpool

* [api] DataCrimeRepository / Data311Repository: refactor findAllByLatLngBoundsAndBetweenDates native SQL to avoid errant N+1 queries with Category associations
remove deprecated findAllByLatLngDistanceAndBetweenDates method
update DataCrimeRepositoryTest

* SourceRepository:findAll() add @EntityGraph annotation to eager load 'locale' and 'mapping' associations

* DataJob: add dervied column for sourceId and getSourceId() method

* DataJobRepository: add findMaxSessionIdOffsetDataJobs native query leveraging GROUP BY source_id to retrieve latest DataJob for all source_id

* DataJobService: replace N+1 dataJobRepository db calls
* use single dataJobRepository.findMaxSessionIdOffsetDataJobs() and in-memory dataJobMap iteration

* WorkerScheduleService: increment scheduled delay to 3500ms

* DataCrime / Data311: refactor id from Integer to Long
66-alter-data-id-to-bigint.sql: migration int to bigint

+++ api/src/main/java/com/quirkshop/nuisancemaps/model/DataEntity.java
@@ -5,9 +5,9 @@ import org.locationtech.jts.geom.Point;
 
 public interface DataEntity {
 
-    public Integer getId();
+    public Long getId();
 
-    public void setId(Integer id);
+    public void setId(Long id);
 
     public String getReportNum();
